fix: move to CSV
Google sheets does not enclose content in a TSV file which means we cannot use multiline markdown to represent a list.  Exporting as a CSV file does enclose content between "", which means the parser used in this app accept `\n` inside a column, which allows us to use multiline markdown.