chore/Add new complaints score field [CP-1059]
From March the existing `complaintsRating` data is changing so that it
can include decimal numbers, which will be rendered to users as the
nearest half star.  At present `complaintsRating`, and the matching
field in Contentful, only include/handle whole numbers.  Once a field
has been added into Contentful there is no way to change it's type.
Therefore a new field has been added into Contentful called.
`complaintsRatingScore`.  The code that populates this field is hidden
behind a feature flag to ensure it's not updated during the December
CSR upload