chore/Update Complaints rating
From March 2026 the complaintRating score will be represented as a decimal, so that scores can be represented as the nearest half star. At present the complaintRating field in Contentful is set as an integer type, which means it can't accept decimals. Therefore a new field has been added into Contentful called complaintsRatingScore. The code that populates this field is hidden behind a feature flag to ensure it's not actioned during the December CSR upload