chore/Add billAccuracyAndMeteringRating data [CP-1051]
From March 2026 there will be a new billAccuracyAndMeteringRating data point in the
CSR spreadsheet.  As such the CSR upload tool needs to be updated to
ensure this new data is mapped to the new Contentful field and uploaded.
This code sits behind FF_NEW_CSR_DATA so it doesn't cause any problems
with the December CSR upload.