chore/Remove FF_NEW_CSR_DATA
The final CSR upload using the original data has been run, which means
the next time it runs in March we need to start permanently populating
the fields in Contentful.  This means that we can remove the feature
flagged code and ensure all the new fields are handled.

The main thing to note is that the new social media sync/async fields
will not all be present for every supplier.  Therefore, if there isn't data present we'll set the sync fields to empty strings and the async fields to null